requirements.txt
ultralytics>=8.1.0
streamlit>=1.31.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0